changed creddit type

+++ StudentManagement/Models/Course.cs
@@ -7,5 +7,5 @@ public class Course
     public int Id { get; set; }
     public string Title { get; set; }
     public Instructor Instructor { get; set; }
-    public float Credits { get; set; }
+    public int Credits { get; set; }
 }
